refactor: Popover의 상태관리 로직 분리 및 openModal 함수 생성
Popover컴포넌트에서 사용하는 상태관리 로직들을 usePopover로 옮기고
click이벤트 함수에서 사용하는 toggleModal, changeModalComponent를 하나로 묶어서 openModal이라고 이름을 지었습니다

Dialog도 기존 Dialog에서 Delete기능만 가지는 DeleteDialog 컴포넌트를 만들어 코드를 좀 더 선언적으로 변경하였습니다
